ADD: Another refactor for custom hooks - useCounter
